Added file exists check to file deleting

+++ src/main/java/de/domisum/lib/auxilium/util/FileUtil.java
@@ -397,7 +397,8 @@ public final class FileUtil
 	{
 		try
 		{
-			Files.delete(file.toPath());
+			if(file.exists())
+				Files.delete(file.toPath());
 		}
 		catch(IOException e)
 		{
